Renamed db.py into db_adapters.py
